+++ consensus/common/src/main/java/org/hyperledger/besu/consensus/common/bft/events/IbftEvent.java
@@ -12,20 +12,9 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.consensus.common.bft.ibftevent;
+package org.hyperledger.besu.consensus.common.bft.events;
 
-import org.hyperledger.besu.ethereum.p2p.rlpx.wire.Message;
-
-/** Static helper functions for producing and working with IbftEvent objects */
-public class IbftEvents {
-  public static IbftEvent fromMessage(final Message message) {
-    return new IbftReceivedMessageEvent(message);
-  }
-
-  public enum Type {
-    ROUND_EXPIRY,
-    NEW_CHAIN_HEAD,
-    BLOCK_TIMER_EXPIRY,
-    MESSAGE
-  }
+/** Category of events that will effect and are interpretable by the Ibft processing mechanism */
+public interface IbftEvent {
+  IbftEvents.Type getType();
 }
