+++ consensus/common/src/main/java/org/hyperledger/besu/consensus/common/bft/events/IbftReceivedMessageEvent.java
@@ -12,9 +12,24 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.consensus.common.bft.ibftevent;
+package org.hyperledger.besu.consensus.common.bft.events;
 
-/** Category of events that will effect and are interpretable by the Ibft processing mechanism */
-public interface IbftEvent {
-  IbftEvents.Type getType();
+import org.hyperledger.besu.ethereum.p2p.rlpx.wire.Message;
+
+public class IbftReceivedMessageEvent implements IbftEvent {
+
+  private final Message message;
+
+  public IbftReceivedMessageEvent(final Message message) {
+    this.message = message;
+  }
+
+  public Message getMessage() {
+    return message;
+  }
+
+  @Override
+  public IbftEvents.Type getType() {
+    return IbftEvents.Type.MESSAGE;
+  }
 }
