+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/IbftEventQueue.java
@@ -14,7 +14,7 @@
  */
 package org.hyperledger.besu.consensus.ibft;
 
-import org.hyperledger.besu.consensus.common.bft.ibftevent.IbftEvent;
+import org.hyperledger.besu.consensus.common.bft.events.IbftEvent;
 
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
