+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/IbftProcessor.java
@@ -15,7 +15,7 @@
 package org.hyperledger.besu.consensus.ibft;
 
 import org.hyperledger.besu.consensus.common.bft.EventMultiplexer;
-import org.hyperledger.besu.consensus.common.bft.ibftevent.IbftEvent;
+import org.hyperledger.besu.consensus.common.bft.events.IbftEvent;
 
 import java.util.Optional;
 import java.util.concurrent.CountDownLatch;
