+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/RoundTimer.java
@@ -15,7 +15,7 @@
 package org.hyperledger.besu.consensus.ibft;
 
 import org.hyperledger.besu.consensus.common.bft.ConsensusRoundIdentifier;
-import org.hyperledger.besu.consensus.common.bft.ibftevent.RoundExpiry;
+import org.hyperledger.besu.consensus.common.bft.events.RoundExpiry;
 
 import java.util.Optional;
 import java.util.concurrent.ScheduledFuture;
