+++ consensus/ibft/src/test/java/org/hyperledger/besu/consensus/ibft/IbftProcessorTest.java
@@ -26,7 +26,7 @@ import static org.mockito.Mockito.verify;
 
 import org.hyperledger.besu.consensus.common.bft.ConsensusRoundIdentifier;
 import org.hyperledger.besu.consensus.common.bft.EventMultiplexer;
-import org.hyperledger.besu.consensus.common.bft.ibftevent.RoundExpiry;
+import org.hyperledger.besu.consensus.common.bft.events.RoundExpiry;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
